deps: make reth and tempo deps optional in `tempo-alloy`
