+++ crates/alloy/src/rpc/mod.rs
@@ -1,3 +1,6 @@
 mod request;
 
 pub use request::{FeeToken, TempoTransactionRequest};
+
+#[cfg(feature = "tempo-compat")]
+mod compat;
